requirements.txt
bs4
requests
xmltodict
prometheus_client
pyyaml
croniter
playwright